requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.0
contourpy==1.1.1
cycler==0.12.1
fonttools==4.43.1
idna==3.4
kiwisolver==1.4.5
matplotlib==3.8.0
numpy==1.26.0
packaging==23.2
Pillow==10.0.1
pyparsing==3.1.1
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
urllib3==2.0.6
python-decouple
mysql
flask
plotly
